requirements.txt
pymongo==4.5.0
requests==2.31.0
python-dotenv==1.0.0
bcrypt==4.0.1  # Para hashear contraseñas (opcional)
matplotlib==3.7.3  # Para visualización (opcional)